:recyle: refactor: 검증 로직을 Validator로 분리 및 매직 넘버를 상수로 대체
